% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd_na.R
\name{average_svd_na}
\alias{average_svd_na}
\title{Average SVD (with missing values)}
\usage{
average_svd_na(Y_list, nu = NULL, nv = NULL)
}
\arguments{
\item{Y_list}{list of matrices}

\item{nu}{number of left singular vectors to compute}

\item{nv}{number of right singular vectors to compute}
}
\description{
Average SVD (with missing values)
}
