% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{remove_spatial}
\alias{remove_spatial}
\title{Normalize feature matrices.}
\usage{
remove_spatial(
  f_mats,
  adj_idx,
  row_rep,
  col_rep,
  d,
  row_impute = TRUE,
  row_center = TRUE,
  verbose = TRUE,
  full.results = FALSE
)
}
\arguments{
\item{f_mats}{feature matrices}

\item{adj_idx}{indices of which features to use to learn spatial effects}

\item{row_rep}{row replicate factor}

\item{col_rep}{column replicate factor}

\item{d}{dimension of spatial effects to remove}

\item{row_impute}{impute missing values in f_mats row-wise as col_rep grouped means}

\item{row_center}{subtract mean from f_mats rows}

\item{verbose}{print intermediate results}

\item{full.results}{return intermediate results. If FALSE just returns adjusted matrices.}
}
\description{
Normalize feature matrices.
}
